Refactor of gateway module
