+++ healthchain/gateway/clients/fhir/__init__.py
@@ -1,0 +1,7 @@
+from .aio.client import AsyncFHIRClient
+from .sync.client import FHIRClient
+
+__all__ = [
+    "AsyncFHIRClient",
+    "FHIRClient",
+]
