+++ healthchain/gateway/clients/fhir/aio/__init__.py
@@ -1,0 +1,3 @@
+from .client import AsyncFHIRClient, create_async_fhir_client
+
+__all__ = ["AsyncFHIRClient", "create_async_fhir_client"]
