+++ healthchain/gateway/clients/fhir/sync/__init__.py
@@ -1,0 +1,3 @@
+from .client import FHIRClient, create_fhir_client
+
+__all__ = ["FHIRClient", "create_fhir_client"]
