+++ healthchain/gateway/fhir/__init__.py
@@ -1,0 +1,12 @@
+from .base import BaseFHIRGateway
+from .aio import AsyncFHIRGateway
+from .sync import FHIRGateway
+from .errors import FHIRErrorHandler, FHIRConnectionError
+
+__all__ = [
+    "BaseFHIRGateway",
+    "AsyncFHIRGateway",
+    "FHIRGateway",
+    "FHIRErrorHandler",
+    "FHIRConnectionError",
+]
