--- healthchain/gateway/protocols/apiprotocol.py
@@ -1,14 +1,0 @@
-from enum import Enum
-
-
-class ApiProtocol(Enum):
-    """
-    Enum defining the supported API protocols.
-
-    Available protocols:
-    - soap: SOAP protocol
-    - rest: REST protocol
-    """
-
-    soap = "SOAP"
-    rest = "REST"
