+++ healthchain/gateway/soap/__init__.py
@@ -1,0 +1,10 @@
+from .notereader import NoteReaderService
+from .utils.epiccds import CDSServices
+from .utils.model import ClientFault, ServerFault
+
+__all__ = [
+    "NoteReaderService",
+    "CDSServices",
+    "ClientFault",
+    "ServerFault",
+]
