+++ healthchain/gateway/soap/utils/wsgi.py
@@ -4,8 +4,8 @@ from spyne.server.wsgi import WsgiApplication
 
 from typing import Callable
 
-from healthchain.gateway.soap.epiccdsservice import CDSServices
-from healthchain.gateway.soap.model import ClientFault, ServerFault
+from healthchain.gateway.soap.utils.epiccds import CDSServices
+from healthchain.gateway.soap.utils.model import ClientFault, ServerFault
 
 
 def start_wsgi(
